Move progress-bar-2 stories to seprate file
